feat(auction): Task T pack Bidder/Position to 2 slots; E1 default=4
uint80 packed structs with overflow guards. Warm ALL-SIMPLE@300 still ~6.09M
(misses 2.5M); maxClearsPerSync default = floor(25M/measured)=4 with
cap*measured<=25M assert. Ladder green under production=lazy.

+++ contracts/test/EpochSync.t.sol
@@ -90,6 +90,6 @@ contract EpochSyncTest is Test {
     function _bid(StonkzAuction a, address who, uint256 budget) internal {
         vm.deal(who, budget + BID_FEE + 1 ether);
         vm.prank(who);
-        a.placeBid{value: budget + BID_FEE}(budget, type(uint256).max);
+        a.placeBid{value: budget + BID_FEE}(budget, type(uint80).max);
     }
 }
